Moved scene manager
