+++ src/game_engine/unity/scene_manager/scene.rs
@@ -1,0 +1,40 @@
+use super::SceneManager;
+use crate::{string::ArrayCString, Address, Error, Process};
+
+/// A scene loaded in the attached game.
+pub struct Scene {
+    pub(super) address: Address,
+}
+
+impl Scene {
+    /// Returns the address of the scene in the attached game.
+    pub const fn address(&self) -> Address {
+        self.address
+    }
+
+    /// Returns [`true`] if the address of the scene still points to valid
+    /// memory.
+    pub fn is_valid(&self, process: &Process) -> bool {
+        process.read::<u8>(self.address).is_ok()
+    }
+
+    /// Returns the build index of the scene. This index is unique to each
+    /// scene in the game.
+    pub fn index(&self, process: &Process, scene_manager: &SceneManager) -> Result<i32, Error> {
+        process.read(self.address + scene_manager.offsets.build_index)
+    }
+
+    /// Returns the full path to the scene.
+    pub fn path<const N: usize>(
+        &self,
+        process: &Process,
+        scene_manager: &SceneManager,
+    ) -> Result<ArrayCString<N>, Error> {
+        process
+            .read_pointer(
+                self.address + scene_manager.offsets.asset_path,
+                scene_manager.pointer_size,
+            )
+            .and_then(|addr| process.read(addr))
+    }
+}
